+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog for Cldr_Units v1.0.2
+# Changelog for Cldr_Units v1.1,0
 
 This is the changelog for Cldr v1.0.2 released on January 12th, 2018.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr_units/tags)
 
@@ -7,6 +7,8 @@ This is the changelog for Cldr v1.0.2 released on January 12th, 2018.  For older
 * Fixes the documentation for `Cldr.Units.convert/2` and adds the correct spec.  Thanks to @lostbokrakai.  Closes #5.
 
 ## Enhancements
+
+* Add `Cldr.Unit.value(Unit.t)` to return the value of the `Unit`
 
 * Format the code with `mix format`